feat(gpu): gpu acceleration supported; updated the dependencies
Along with that corrected the defaults according to the local systems'
poor capabilities.

Updated the dependencies - reduced the amount of packages in favour
of only those used.

Divided the requirements into two files - one for the production and one
for development.

Written a separate script to build a .whl file with appropriate
qiskit-aer version for used system.