test(config): refactored the structure, provided tests for the config classes
Removed the *_config suffix from the config modules and moved them to
quantum_pipeline/configs/module/ directory. Thanks to that it looks
a bit nicer ex.:

from quantum_pipeline.configs.module.backend import BackendConfig

Along with that provided tests for:
  - BackendConfig
  - ProducerConfig
  - SecurityConfig (along with other security classes - CertConfig etc.)
  - ConfigurationManager

The rest of edits is just a refactor of the imports in the codebase
along with slight improvements to existing tests (including automatic
skip for kafka tests if the broker is not available).